Add files via upload
In this project we will learn Pose Estimation. We will detect 33 difference landmarks within a human body and all of this will be done in real time. That’s right more than 24 fps. We will first look into the basic code required to run and then create a module out of this so that we don’t have to write the code again and again for different projects